requirements.txt
certifi==2020.6.20
cffi==1.14.3
cryptography==3.1.1
decorator==4.4.2
numpy==1.19.2
opencv-python==4.4.0.44
pycparser==2.20
PySocks==1.7.1
python-telegram-bot==12.8
six~=1.11.0
tornado==6.0.4
pip~=20.1.1
configparser~=3.8.1
pytz~=2020.1
uuid~=1.30
setuptools~=47.1.0
pandas~=1.1.2
python-dateutil~=2.8.1
colorama~=0.4.3
Jinja2~=2.11.2
click~=7.1.2
MarkupSafe~=1.1.1
scipy~=1.5.2
sklearn~=0.0
scikit-learn~=0.23.2
pypiwin32~=223
pywin32~=228
future~=0.18.2
joblib~=0.17.0
threadpoolctl~=2.1.0